all the info about anything or anyone
This software can help you with anything you can desire. It can help with schoolwork like math, science, and even history. Not just US history, the entire world's history. It can help learn another language. It can help you even learn another culture. It can help you learn landmarks like the Eiffel Tower, Leaning Tower of Pisa, and even the Great Wall of China. But the most important thing is this is not for copying, this is just for understanding about it.